Attempts to implement an automatic early stop of the nn. Initially wanted to use an API similar to keras, where an object which could handle this was implemented. The module I found seems promising, but there are some errors wrt. indexing of some 1d tensors which produces erors. If these are corrected in the source code for the module, then it works, but this requires a lot of work for anyone who would want to use the code on another machine.
I have attempted to set up a mechanism in my own implementation
mainly based on recording each increase in accuracy and noting
whenever it dips. If it then stays bellow this best accuracy for
a number of epochs I break the loop. This seems sound to me, but
it doesn't seem to be working. hanging up the towel for now.